+++ main/output.c
@@ -715,7 +715,7 @@ static int php_ub_body_write(const char *str, uint str_length TSRMLS_DC)
 PHP_FUNCTION(ob_start)
 {
 	zval *output_handler=NULL;
-	uint chunk_size=0;
+	long chunk_size=0;
 	zend_bool erase=1;
 	int argc = ZEND_NUM_ARGS();
